fix: [CDS-59320]: changes in additional variable not reflecting in service v2
